Make a functional datastore that fastapi can use
FastAPI can now send the incoming data to the datastore where it will be
stored in a dataframe. Incoming data is not immediately put into the
dataframe, as this generally causes copies for each incoming line. It is
instead added to a queue that will be merged with the dataframe when the
dataframe is read from.