Move tests to use dependency injection and test_archive
We can construct a datastore dependency with the smaller archive,
speeding up testing. app.dependency_overrides is a dict, the keys are
the callables that are used as dependencies in the app and the values
are the overrides. To remove an override, remove it and the key from the
dict. To remove all overrides, set dependency_overrides to an empty
dict.