Use Annotated to package functionality with Format enum
Rather than match or if else on the enum in the code, the format enum
itself should know what needs to be done in each situation. This wraps
serializing and deserializing functions, as well as endpoint information
for the different formats.

It keeps the typechecker mostly happy, but it isn't perfect. It's
somewhat boilerplatey to set up, but presents a simple interface to use.
All formats must provide all the required annotations, so it isn't very
flexible, but works for this case.