Get rid of enum in favor of protocol implementation
The enum was weird and implemented behaviour specific to a class it
didn't relate to. It wasn't a member or in the inheritance hierarchy.

This gets rid of that in favor of a Protocol inheritance hierarchy. An
overarching SerializableFormat with static methods.
This protocol is inherited by two more specialized protocols that know
more specific types when dealing with Parquet and JSON. JSON serializes
to strings, while Parquet serializes to bytes. A dataframe in general
can be serialized to strings, bytes, or even buffers. This hierarchy
allows us to express this in the type system.

Actual implementation of the protocols should be done in classes next to
schemas that then own the implementations. They know how to handle
things like JSON orient= and the index in parquet for their schema.

No other schemas are planned, but this allows us to extend with new ones
if we need to.