Utility functions for determining date ranges for given week
The plan is to save a parquet file for each ISO 8601-week. A specific week
would be identified by the ISO year and the week number.

In order to determine what data belongs where the easiest solution is to
simply find the datetime point the week starts and the next week starts.

Useful behaviour of ISO is that Jan 4th is always in week 1, and Dec
28th is always in the last week, although that week may be number 52 or
53.

For the end we advance 7 days rather than 1 ISO week, as the ISO week
may wrap on the year.