Implement a manager for the weekly parquet file structure
A new ParquetManager class handles loading and saving the data that is
now stored in weekly chunks. It itself holds no state apart from the
location of the archives, which will be used for testing.

This manager will load the data for the current ongoing week, and the
previous week. This ensures at least 7 days of data is available.

When the datastore archives its data it will also reload it from disk
afterwards, so it doesn't keep data from more than 2 weeks ago in
memory.

If the server is asked for data older than is loaded in memory, the
ParquetManager will load that from disk into a dataframe. That dataframe
will then be discarded after it is serialized.