Make a custom handler for the latest incoming data
Appending the latest incoming data to the dataframe is expensive. When a
plotter is attached, the data is always appended to the dataframe when a
request is made, to keep it in sync. This causes a heavy load on the
server, and can be observed to slow down data sampling on the
sensor-client.

The aim of this class is to be a quick access to the latest data that we
can generate a json dataframe from without concatenating it to the main
dataframe.

It uses a deque in its implementation, as it can be used to quickly
append data, and it will rotate old data out automatically. We know the
latest data is always at the front, but we will need to be careful to
keep a check on which data has not yet been synced.