Improve code to pass tests
Various fixes to the code to pass the tests.

Moves parquetmanager to its own file and puts actual saving into a
separate "backend" class. One backend saves parquet-files to disk, and
another just keeps the dataframes in memory in a python dictionary. This
dict-backend is used in the testing-framework and provides much much
faster testing.

The code should now be more robust to data that is currently illogical
and out-of-order, since hypothesis is generating some data that is not
expected to exist normally.

SensorData.repair_dataframe now handles more things that can be wrong
with the frame.