Fix failing test case where old data was overwritten
The tests assume that in the case of data with duplicated index (exact
same sensor_type, sensor and timestamp), only 1 entry is saved, and it's
the first one given.

When merging, the order got flipped. The dataframe that was created had
the quick_df as it's "oldest" or "first" source of data, so if there was
a duplicate between the quick and in_memory df, the in_memory df entry
got dropped. This happened in testing. The fix is therefore to properly
order the dataframes by 'seniority'

In the real world this should never happen, so how
we deal with it is rather arbitrary. If it happens, the backend will
drop the data, and with this behaviour, will keep the first data it got,
and not report any errors.

Error-reporting might be appropriate.